# server/openapi.yaml
openapi: 3.0.3
info:
  title: Media Management System API
  description: |
    Backend API for a comprehensive media management system supporting images, social media content, and video reels.
    
    ## Features
    - Card management (image, social, reel types)
    - File uploads with automatic processing
    - User authentication and management
    - Password reset via email (SendGrid or Mailgun)
    - Tag management
    - Health monitoring with dependency checks
    - Both local and S3 storage support
    - Soft delete system with trash management (admin/editor only)
    - Centralized error handling with correlation ID tracking
    - Database connection pooling with retry logic
    - Environment validation on startup
    - Secure system initialization with setup endpoint
    
    ## Architecture Improvements
    - **Correlation IDs**: All responses include `X-Correlation-ID` header for request tracking
    - **Centralized Error Handling**: Consistent error responses with correlation IDs and timestamps
    - **Enhanced Health Checks**: Comprehensive monitoring of database, storage, email, and system resources
    - **Database Resilience**: Connection pooling, retry logic with exponential backoff
    - **Structured Logging**: Correlation ID tracking throughout the request lifecycle
    - **Environment Validation**: Startup validation of all environment variables
    
    ## Authentication
...
    
    ## File Upload
...
    
    ## Email Configuration
...
    - Mailgun
    
...
  contact:
...
servers:
...
paths:
...
  /api/cards/{id}/download-package:
    get:
      tags:
        - Files
      summary: Create download package
      description: Create a ZIP package containing all files for a card
      parameters:
        - name: id
          in: path
          required: true
          description: Card ID
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: Download URL created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadPackageResponse'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/download-zip/{filename}:
    get:
      tags:
        - Files
      summary: Download ZIP package
      description: Download a previously created ZIP package
      parameters:
        - name: filename
          in: path
          required: true
          description: ZIP filename
          schema:
            type: string
      responses:
        '200':
          description: ZIP file download
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/files/signed-url:
    post:
      tags:
        - Files
      summary: Generate signed URL
      description: Generate a signed URL for S3 files (only works when S3 is configured)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignedUrlRequest'
      responses:
        '200':
          description: Signed URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedUrlResponse'
        '400':
          description: Invalid URL or S3 not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Tag Management
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
  schemas:
...
    DownloadPackageResponse:
      type: object
      properties:
        downloadUrl:
          type: string
          description: URL to download ZIP package
          example: "/api/download-zip/card_123_files.zip"
    SignedUrlRequest:
      type: object
      required:
        - fileUrl
      properties:
        fileUrl:
          type: string
          format: uri
          description: S3 file URL to generate signed URL for
          example: "https://bucket.s3.amazonaws.com/uploads/file.jpg"
    SignedUrlResponse:
      type: object
      properties:
        signedUrl:
          type: string
          format: uri
          description: Temporary signed URL for file access
          example: "https://bucket.s3.amazonaws.com/uploads/file.jpg?X-Amz-Signature=..."

    # User Schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request parameters"
        details:
          type: string
          description: Additional error details
          example: "Username is required"
        correlationId:
          type: string
          format: uuid
          description: Unique request correlation ID for tracking
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
        path:
          type: string
          description: Request path where error occurred
        method:
          type: string
          description: HTTP method where error occurred
...